Revert "Revert "Merge pull request #26859 from vespa-engine/revert-26805-olaa/versioned-feature-flag""
This reverts commit e10612188741a2d5ce9c69b80c35015b38ea1627.
