+++ container-disc/src/main/java/com/yahoo/container/jdisc/AthenzIdentityProviderProvider.java
@@ -89,9 +89,6 @@ public class AthenzIdentityProviderProvider implements Provider<AthenzIdentityPr
         public Path trustStorePath() {
             throw new UnsupportedOperationException(message);
         }
-
-        @Override
-        public void deconstruct() {}
     }
 
 }
