+++ flags/src/main/java/com/yahoo/vespa/flags/Flags.java
@@ -383,7 +383,7 @@ public class Flags {
             List.of("olaa"), "2023-04-12", "2023-06-12",
             "Whether AthenzCredentialsMaintainer in node-admin should create tenant service identity certificate",
             "Takes effect on next tick",
-            ZONE_ID, HOSTNAME, VESPA_VERSION, APPLICATION_ID
+            ZONE_ID, HOSTNAME
     );
 
     public static final UnboundBooleanFlag ENABLE_CROWDSTRIKE = defineFeatureFlag(
